requirements.txt

fastapi
uvicorn
crewai
litellm
python-dotenv
supabase
